// src/app/modules/login/services/auth/types/EPermission.ts
export enum EPermission {
	APIAccess = 'APIAccess',
	CanUserLoginToPlace = 'CanUserLoginToPlace',
	CurrencyRead = 'CurrencyRead',
	CurrencyCreate = 'CurrencyCreate',
	CurrencyEdit = 'CurrencyEdit',
	CurrencyDelete = 'CurrencyDelete',
	CurrenciesAccountsRead = 'CurrenciesAccountsRead',
	CurrenciesAccountsEdit = 'CurrenciesAccountsEdit',
	CardRead = 'CardRead',
	CardAssign = 'CardAssign',
	CardEdit = 'CardEdit',
	CardBlock = 'CardBlock',
	CardUnBlock = 'CardUnBlock',
	CardDelete = 'CardDelete',
	CardReadUser = 'CardReadUser',
	GoodsTypeRead = 'GoodsTypeRead',
	GoodsTypeCreate = 'GoodsTypeCreate',
	GoodsTypeEdit = 'GoodsTypeEdit',
	GoodsTypeDelete = 'GoodsTypeDelete',
	GoodsRead = 'GoodsRead',
	GoodsReadOwn = 'GoodsReadOwn',
	GoodsCreate = 'GoodsCreate',
	GoodsCreateOwn = 'GoodsCreateOwn',
	GoodsEdit = 'GoodsEdit',
	GoodsEditOwn = 'GoodsEditOwn',
	GoodsDelete = 'GoodsDelete',
	UserRead = 'UserRead',
	UserReadOwn = 'UserReadOwn',
	UserCreate = 'UserCreate',
	UserEdit = 'UserEdit',
	UserDelete = 'UserDelete',
	PlaceRead = 'PlaceRead',
	PlaceReadDetail = 'PlaceReadDetail',
	PlaceReadOwn = 'PlaceReadOwn',
	PlaceCreate = 'PlaceCreate',
	PlaceEdit = 'PlaceEdit',
	PlaceEditOwnGoods = 'PlaceEditOwnGoods',
	PlaceDelete = 'PlaceDelete',
	TransactionRead = 'TransactionRead',
	TransactionManage = 'TransactionManage',
	TransactionCreatePayment = 'TransactionCreatePayment',
	TransactionCreateDeposit = 'TransactionCreateDeposit',
	TransactionCreateWithdraw = 'TransactionCreateWithdraw',
	TransactionCancellation = 'TransactionCancellation',
	StatisticsRead = 'StatisticsRead',
	LogsRead = 'LogsRead',
}